Restructure format tests into separate files
