+++ test/formats/test_aac.py
@@ -1,0 +1,10 @@
+from .common import CommonTests
+
+
+class AACTest(CommonTests.SimpleFormatsTest):
+    testfile = 'test.aac'
+
+
+class AACWithAPETest(CommonTests.TagFormatsTest):
+    testfile = 'test-apev2.aac'
+    supports_ratings = False
