+++ test/formats/test_asf.py
@@ -1,0 +1,11 @@
+from .common import CommonTests
+
+
+class ASFTest(CommonTests.TagFormatsTest):
+    testfile = 'test.asf'
+    supports_ratings = True
+
+
+class WMATest(CommonTests.TagFormatsTest):
+    testfile = 'test.wma'
+    supports_ratings = True
