+++ test/formats/test_midi.py
@@ -1,0 +1,5 @@
+from .common import CommonTests
+
+
+class MIDITest(CommonTests.SimpleFormatsTest):
+    testfile = 'test.mid'
