+++ test/formats/test_wav.py
@@ -1,0 +1,5 @@
+from .common import  CommonTests
+
+
+class WAVTest(CommonTests.SimpleFormatsTest):
+    testfile = 'test.wav'
